A way to retrieve the canonical part of constant and mind out of the user part
For that, in the Pre_env map env_constants, we add a copy of the full constant
name in the data part of the Cmap_env.t. Since these maps are searched according
to the user part only, we can use make_con to create a first constant (whose
canonical part might be wrong), and then search for the complete constant name.
Cf new function:
 Environ.make_con_form_user : env -> module_path -> label -> constant

Same for mutual_inductive:
 Environ.make_mind_form_user : env -> module_path -> label -> mutual_inductive